// src/app/modules/kpi/data-access/models/kpi-managers/indicators.model.ts
import { FileAttach } from '@shared/model/file-attach.model';


export class IndicatorsModel {

  indicatorId?: number;
  name?: string;
  unitId?: string;
  periodType?: string;
  significance?: string;
  measurement?: string;
  systemInfo?: string;
  type?: string;
  note?: string;
  listValues?: string;
  ratingType?: string;
  docIdsDelete?: number[];
}